+++ .storybook/config.js
@@ -11,9 +11,20 @@ addDecorator(withPropsTable)
 addDecorator(CssResetWrapper)
 addDecorator(FrameStylesWrapper)
 
+const includeTesting = 'STORYBOOK_INCLUDE_TESTING' in process.env
+
 function loadStories() {
     const req = require.context('../stories', true, /\.stories\.js$/)
     req.keys().forEach(filename => req(filename))
 }
 
-configure(loadStories, module)
+function loadStoriesInclTesting() {
+    const req = require.context(
+        '../stories',
+        true,
+        /\.stories(\.testing)?\.js$/
+    )
+    req.keys().forEach(filename => req(filename))
+}
+
+configure(includeTesting ? loadStoriesInclTesting : loadStories, module)