+++ cypress/integration/HeaderBar/LogoHomeLink/index.js
@@ -1,0 +1,11 @@
+import '../common/index.js'
+
+Then('the HeaderBar should display the dhis2 logo', () => {
+    cy.get('[data-test-id="headerbar-logo"]').should('be.visible')
+})
+
+Then('the logo should link to the homepage', () => {
+    cy.get('[data-test-id="headerbar-logo"] a').then($a => {
+        expect($a.attr('href')).to.equal('/')
+    })
+})